+++ .vscode/c_cpp_properties.json
@@ -1,40 +1,58 @@
 {
-  "configurations": [
-    {
-      "name": "Win32",
-      "cStandard": "c17",
-      "cppStandard": "c++17",
-      "intelliSenseMode": "windows-msvc-x64",
-      "defines": [
-        "NAPI_DISABLE_CPP_EXCEPTIONS"
-      ],
-      "includePath": [
-        "C:/Users/${env:username}/AppData/Local/node-gyp/Cache/18.16.0/include/node",
-        "${workspaceFolder}/node_modules/node-addon-api",
-        "${workspaceFolder}/src/addon",
-        "${workspaceFolder}/src/addon/tools/src/wincompat"
-      ]
-    },
-    {
-      "name": "Linux",
-      "compilerPath": "/usr/bin/clang",
-      "cStandard": "c17",
-      "cppStandard": "c++17",
-      "intelliSenseMode": "linux-clang-x64",
-      "compilerArgs": [
-        "-fpermissive"
-      ],
-      "defines": [
-        "NAPI_DISABLE_CPP_EXCEPTIONS"
-      ],
-      "includePath": [
-        "${workspaceFolder}/node_modules/node-addon-api/**",
-        "${workspaceFolder}/node_modules/**",
-        "${workspaceFolder}/**",
-        "/usr/include/node",
-        "${workspaceFolder}/src/addon"
-      ]
-    }
-  ],
-  "version": 4
+    "configurations": [
+        {
+            "name": "Win32",
+            "cStandard": "c17",
+            "cppStandard": "c++17",
+            "intelliSenseMode": "windows-msvc-x64",
+            "defines": [
+                "NAPI_DISABLE_CPP_EXCEPTIONS"
+            ],
+            "includePath": [
+                "C:/Users/${env:username}/AppData/Local/node-gyp/Cache/18.16.0/include/node",
+                "${workspaceFolder}/node_modules/node-addon-api",
+                "${workspaceFolder}/src/addon",
+                "${workspaceFolder}/src/addon/tools/src/wincompat"
+            ]
+        },
+        {
+            "name": "Linux",
+            "compilerPath": "/usr/bin/clang",
+            "cStandard": "c17",
+            "cppStandard": "c++17",
+            "intelliSenseMode": "linux-clang-x64",
+            "compilerArgs": [
+                "-fpermissive"
+            ],
+            "defines": [
+                "NAPI_DISABLE_CPP_EXCEPTIONS"
+            ],
+            "includePath": [
+                "${workspaceFolder}/node_modules/node-addon-api/**",
+                "${workspaceFolder}/node_modules/**",
+                "${workspaceFolder}/**",
+                "/usr/include/node",
+                "${workspaceFolder}/src/addon"
+              ]
+            },
+            {
+              "name": "Mac",
+              "includePath": [
+                "/usr/local/include/node",
+                "${workspaceFolder}/node_modules/node-addon-api",
+                "${workspaceFolder}/src/addon/**"
+              ],
+            "defines": [
+                "NAPI_DISABLE_CPP_EXCEPTIONS"
+            ],
+            "macFrameworkPath": [
+                "/Library/Developer/CommandLineTools/SDKs/MacOSX.sdk/System/Library/Frameworks"
+            ],
+            "compilerPath": "/usr/bin/clang",
+            "cStandard": "c17",
+            "cppStandard": "c++17",
+            "intelliSenseMode": "macos-clang-x64"
+        }
+    ],
+    "version": 4
 }
