+++ package.json
@@ -41,13 +41,13 @@
   "devDependencies": {
     "@types/chai": "^4.3.5",
     "@types/mocha": "^10.0.1",
-    "@types/node": "^20.1.3",
+    "@types/node": "^20.4.2",
     "mocha": "^10.2.0",
-    "node-gyp": "^9.3.1",
+    "node-gyp": "^9.4.0",
     "ts-node": "^10.9.1",
-    "typescript": "^5.0.4"
+    "typescript": "^5.1.6"
   },
   "dependencies": {
-    "node-addon-api": "^6.1.0"
+    "node-addon-api": "^7.0.0"
   }
 }
